Decoupled Model, Controller and View
Decoupled the Controller from the Model and the View. Controllers, models, and views mostly only know about their abstract interfaces. 

Controller now takes a concrete implementation of a model and a view during initialization. Besides the base Model and View classes that guarantee their services, for this example there is a ProductModel and a ConsoleView to serve as their concrete implementations.